ci: add release workflow for github actions
This workflow will be triggered when a push is performed to the
main branch. It will check if there are any changes
in the .changeset folder.

If so, it will create a new pull request that will be updated with
the new changes added. If the PR is approved, a new release of the
changed packages will be made.

Otherwise, all packages that have not been released to npm will be
released by default.